servers/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Mail
Flask-Migrate
Flask-CORS
Flask-SocketIO
psycopg2-binary
pymongo
redis
python-dotenv
PyJWT
spacy
scikit-learn
torch
transformers
PyPDF2
python-docx
cloudinary
celery
gunicorn